move some of the contract structs to primitives
