+++ chain-signatures/contract/src/primitives.rs
@@ -1,10 +1,47 @@
+use crypto_shared::{derive_epsilon, SerializableScalar};
+use k256::Scalar;
 use near_sdk::borsh::{self, BorshDeserialize, BorshSerialize};
 use near_sdk::serde::{Deserialize, Serialize};
-use near_sdk::{AccountId, PublicKey};
+use near_sdk::{AccountId, BorshStorageKey, CryptoHash, PublicKey};
 use std::collections::{BTreeMap, HashSet};
 
 pub mod hpke {
     pub type PublicKey = [u8; 32];
+}
+
+#[derive(BorshSerialize, BorshDeserialize, BorshStorageKey, Hash, Clone, Debug, PartialEq, Eq)]
+#[borsh(crate = "near_sdk::borsh")]
+pub enum StorageKey {
+    PendingRequests,
+}
+
+/// The index into calling the YieldResume feature of NEAR. This will allow to resume
+/// a yield call after the contract has been called back via this index.
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug, Clone)]
+#[borsh(crate = "near_sdk::borsh")]
+pub struct YieldIndex {
+    pub data_id: CryptoHash,
+}
+
+#[derive(BorshDeserialize, BorshSerialize, Serialize, Deserialize, Debug, Clone)]
+#[borsh(crate = "near_sdk::borsh")]
+pub struct SignatureRequest {
+    pub epsilon: SerializableScalar,
+    pub payload_hash: SerializableScalar,
+}
+
+impl SignatureRequest {
+    pub fn new(payload_hash: Scalar, predecessor_id: &AccountId, path: &str) -> Self {
+        let epsilon = derive_epsilon(predecessor_id, path);
+        let epsilon = SerializableScalar { scalar: epsilon };
+        let payload_hash = SerializableScalar {
+            scalar: payload_hash,
+        };
+        SignatureRequest {
+            epsilon,
+            payload_hash,
+        }
+    }
 }
 
 #[derive(
